+++ dapper/app/load_app.go
@@ -14,7 +14,7 @@ import (
 	"encoding/json"
 	"io/ioutil"
 
-	dt "github.com/singulatron/singulatron/localtron/dapper/types"
+	dt "github.com/singulatron/singulatron/dapper/types"
 )
 
 // LoadAppConfiguration loads an application configuration from a JSON file.
