+++ dapper/app/run.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 	"time"
 
-	dt "github.com/singulatron/singulatron/localtron/dapper/types"
+	dt "github.com/singulatron/singulatron/dapper/types"
 )
 
 func (cm *ConfigurationManager) Run(app *dt.App, tlv map[string]string, anon bool) (*dt.RunContext, error) {
