+++ dapper/features/virtual_machine_platforms.go
@@ -10,7 +10,7 @@
  */
 package features
 
-import dt "github.com/singulatron/singulatron/localtron/dapper/types"
+import dt "github.com/singulatron/singulatron/dapper/types"
 
 var VirtualMachinePlatformFeature = dt.Feature{
 	ID:   "virtual-machine-platform-enabled",
