+++ dapper/features/wsl_installed.go
@@ -10,7 +10,7 @@
  */
 package features
 
-import dt "github.com/singulatron/singulatron/localtron/dapper/types"
+import dt "github.com/singulatron/singulatron/dapper/types"
 
 var WslInstalled = dt.Feature{
 	ID:   "wsl-installed",
