+++ dapper/features/wsl_updated.go
@@ -10,32 +10,40 @@
  */
 package features
 
-import (
-	dt "github.com/singulatron/singulatron/localtron/dapper/types"
-)
+import dt "github.com/singulatron/singulatron/dapper/types"
 
-var WslEnabled = dt.Feature{
-	ID:   "wsl-enabled",
-	Name: "WSL Enabled",
+var WslUpdated = dt.Feature{
+	ID:   "wsl-updated",
+	Name: "WSL Updated",
+	Arguments: []dt.Argument{
+		{
+			Name:    "wslVersion",
+			Type:    dt.Int,
+			Default: 2,
+		},
+	},
 	PlatformScripts: map[dt.Platform]*dt.Scripts{
 		dt.Windows: {
 			Execute: &dt.Script{
-				Source:  `dism.exe /online /enable-feature /featurename:Microsoft-Windows-Subsystem-Linux /all /norestart`,
+				Source: `
+Write-Host "Updating WSL kernel"
+wsl --update
+`,
 				Runtime: "powershell",
 			},
 			Check: &dt.Script{
 				Source: `
-$wslFeature = dism.exe /online /get-featureinfo /featurename:Microsoft-Windows-Subsystem-Linux
-if ($wslFeature.Contains("State : Enabled")) {
-    return $true
+$wslVersion = wsl --status
+if ($wslVersion.Contains("Default Version: {{.wslVersion}}")) {
+    exit 0
 } else {
-    return $false
+    exit 1
 }`,
 				Runtime: "powershell",
 			},
 		},
 	},
 	PlatformFeatures: map[dt.Platform][]any{
-		dt.Windows: {VirtualMachinePlatformFeature.ID},
+		dt.Windows: {WslInstalled.ID},
 	},
 }
