+++ go.work
@@ -1,3 +1,4 @@
 go 1.21
 
 use ./localtron
+use ./dapper
